pyproject.toml
[tool.black]
include = "\\.pyi?$"
line-length = 120
target-version = ["py36"]

[tool.poetry]
name = "hare"
version = "0.1.1"
description = "Hare: Smart Shortcut Engine."
authors = ["Conveen"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/conveen/hare"
classifiers = [
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
packages = [
    { include = "hare/api" },
    { include = "hare/core" },
    { include = "hare/ui" },
]
include = [
    "hare/ui/static/ui/*.ico",
    "hare/ui/static/ui/**/*.js",
    "hare/ui/static/ui/**/*.css",
    "hare/ui/templates/ui/**/*.html",
]

[tool.poetry.dependencies]
python = ">= 3.6, < 3.10"
Django = "^3.2"
djangorestframework = "^3"
djangorestframework-stubs = "1.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
django-stubs = "^1.8.0"
httpie = "*"
mypy = "^0.812"
pylint = "^2.8.0"
pylint-django = "^2.4.4"

[tool.poetry.extras]
postgres = ["psycopg2"]
postgresb = ["psycopg2-binary"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
